+++ examples/llvm-flake-parts/flake.nix
@@ -1,3 +1,4 @@
+# file: examples/llvm-flake-parts/flake.nix
 {
   inputs = {
     nixpkgs.url = "github:cachix/devenv-nixpkgs/rolling";
@@ -48,5 +49,6 @@
           ];
         };
       };
-    }; # outputs
+    };
+  # outputs
 }